Top Bar OJ - Add Optimizely to root layout
