+++ src/app/components/PageLayoutWrapper/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { render } from '../react-testing-library-with-providers';
-import PageLayoutWrapper from '.';
+import PageLayoutWrapper from './PageLayoutWrapper';
 
 global.performance.getEntriesByName = jest.fn(() => []);
