+++ src/app/pages/ArticlePage/index.tsx
@@ -1,7 +1,4 @@
-import withOptimizelyProvider from '#app/legacy/containers/PageHandlers/withOptimizelyProvider';
 import ArticlePage from './ArticlePage';
 import applyBasicPageHandlers from '../utils/applyBasicPageHandlers';
 
-const ArticlePageWithOptimizely = withOptimizelyProvider(ArticlePage);
-
-export default applyBasicPageHandlers(ArticlePageWithOptimizely);
+export default applyBasicPageHandlers(ArticlePage);
