+++ ws-nextjs-app/pages/[service]/live/[id]/live.stories.tsx
@@ -22,7 +22,6 @@ const mockPageData = {
 };
 
 const Component = ({ pageData }: ComponentProps) => (
-  // @ts-expect-error partial data required for storybook
   <PageLayoutWrapper pageData={pageData} status={200}>
     <Live pageData={pageData} />
   </PageLayoutWrapper>
